requirements.txt
Flask==3.0.3
flask-paginate==2024.4.12
bcrypt==4.1.3
mysql-connector-python==8.4.0
python-dotenv==1.0.1
redis==5.0.8
flask-redis==0.4.0
pymongo==4.8.0